// lib/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          avatar_url: string | null
          full_name: string | null
          id: string
          updated_at: string | null
          username: string | null
        }
        Insert: {
          avatar_url?: string | null
          full_name?: string | null
          id?: string
          updated_at?: string | null
          username?: string | null
        }
        Update: {
          avatar_url?: string | null
          full_name?: string | null
          id?: string
          updated_at?: string | null
          username?: string | null
        }
        Relationships: []
      }
// ...
      video: {
        Row: {
          created_at: string | null
          description: string | null
          dislikes: number | null
          id: string
          likes: number | null
          thumbnailUrl: string | null
          title: string | null
          user_id: string | null
          videoStatus: boolean | null
          videoUrl: string | null
          viewCount: number | null
        }
        Insert: {
          created_at?: string | null
          description?: string | null
          dislikes?: number | null
          id?: string
          likes?: number | null
          thumbnailUrl?: string | null
          title?: string | null
          user_id?: string | null
          videoStatus?: boolean | null
          videoUrl?: string | null
          viewCount?: number | null
        }
        Update: {
          created_at?: string | null
          description?: string | null
          dislikes?: number | null
          id?: string
          likes?: number | null
          thumbnailUrl?: string | null
          title?: string | null
          user_id?: string | null
          videoStatus?: boolean | null
          videoUrl?: string | null
          viewCount?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "video_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}